refactor(ci): split the gate's integrity tests into their own suite
Round 8's tests took scripts/sentry-suite-gate.test.mjs to 1,166 lines, past
the 1,000-line hard cap the checker's file-size pin now enforces on it. The
repo's own remedy, named in that assertion, is a focused sibling module.

Behaviour-neutral, and provably so: 42 cases before, 32 + 10 after, and the
sorted set of case names is byte-identical across the split. No case was
renamed, added, lost, or rewritten — only moved.

  scripts/sentry-suite-gate-integrity.test.mjs (10) — claims about a RUN being
  trustworthy as a whole: digest verification before each spawn and after the
  last, the watch set equalling the derived inputs, re-enumeration catching a
  suite created mid-run, exemption route evidence, and step-summary isolation.

  scripts/sentry-suite-gate.test.mjs (32) — claims about parsing and judging a
  single suite's output: reporters, floors, set equality, the env latch.

  scripts/sentry-suite-gate-fixtures.mjs — the shared fixture helpers. Extracted
  rather than duplicated because `runGate` is now load-bearing (it is what keeps
  fixture gates out of the real $GITHUB_STEP_SUMMARY) and two copies could drift
  apart silently, which is the bug class these suites exist to catch. Not named
  `sentry-*.test.mjs`, so `findSentrySuites` does not enumerate a file with no
  cases.

The split file is itself manifest-enumerated, so the gate runs and floors it
like any other suite: the cap stops being a rule the file merely satisfies and
becomes something the gate enforces. Manifest housekeeping per the equal-the-
count rule: new entry at floor 10, and the original re-measured DOWN from 42 to
32 rather than left sitting above the reduced file. The #1721 coverage rule also
requires a direct `scripts`-job invocation for any scripts/sentry-*.test.mjs, so
ci.yml gains one; the local gate's generic glob already routes it with no new
arm. All four gate files are in the file-size pin.

Refs #1779.
