+++ scripts/check-sentry-suites-in-ci.test.mjs
@@ -977,12 +977,12 @@ test("the checker's own files stay under the file-size hard cap", () => {
     "scripts/check-sentry-suites-in-ci-probes.mjs",
     "scripts/check-sentry-suites-in-ci-lifecycle.test.mjs",
     "scripts/check-sentry-suites-in-ci-gate-job.test.mjs",
-    // The runtime gate and its own suite. They were outside this list while
-    // growing to 833 and 935 lines, so the cap they are subject to was not
-    // actually enforced on them — the same unwatched-drift class this whole
-    // check exists to catch, in the checker's own house.
+    // The runtime gate and its suites: unwatched, they grew to 833 and 935
+    // lines with the cap unenforced — this check's own drift class, at home.
     "scripts/sentry-suite-gate.mjs",
     "scripts/sentry-suite-gate.test.mjs",
+    "scripts/sentry-suite-gate-integrity.test.mjs",
+    "scripts/sentry-suite-gate-fixtures.mjs",
   ];
   const over = files
     .map((file) => ({
